Move ResponseFuture and Worker to dedicted mods
